Update player movement to be more efficient
There was a pretty big bug that drastically changed the speed of the player in different refresh rates.

In 165hz (my native refresh rate), the player.speed was set at 10000 and moved moderately slow. Whilst in 60hz (which is more common), the player was moving extraordinarily fast.

This was because of the way I coded in the equation of player movement.

The original equation for player movement was

player.speed * dt

This was incredibly inefficient since the player.speed was linked to the tickrate of the game and having lower refresh rates made the player incredibly fast.

The equation for Delta Time is:
1/[refresh rate]

So, if you put the equation into the calculator and make the refresh rate 165, it would output this number:
0.0060606...

And if you replace the refresh rate to 60 it would output this number:
0.0166666...

Now, you can obviously see the problem just by looking at both of the numbers on different refresh rates.

If you put this equation to a calculator:
10000 * (1/165)

It would output:
60.606060606...

And if you put this equation into the calculator:
10000 * (1/60)

It would output:
166.6666666...

Hence, giving the player a much bigger speed.

The way I resolved this is by removing the dt completely and instead adding a new variable alongside player.speed which is player.speed_modifier.

The initial value of this variable is 1 and by doing functions like holding shift to sprint it would change to a number like 2.

And the new equation set to player movement is:

player.speed * player.speed_modifier